Python/games/intro/pyproject.toml
[project]
name = "intro"
version = "0.1.0"
description = "Pygame tryout with offical documentation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.5.7",
]
